moving enum to its own file
