+++ Assets/Mirage/Runtime/ClientStoppedReason.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7eb51078953dab4c93fbf759747aac7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
